Precision Changes, Path* interface
Float precisions changed to double, interface added for passing Path*

+++ gl/glpoipoint.cpp
@@ -18,22 +18,22 @@ void GLPOIPoint::setupVertices()
 {
     vertices.clear();
     boost_xy_point geoPos = point->getGeoPosition();
-    vertices.append(QVector3D((geoPos.x() * 100.0f) + 0.1f, (geoPos.y() * 100.0f) + 0.1f, 0.3f));
-    vertices.append(QVector3D((geoPos.x() * 100.0f) - 0.1f, (geoPos.y() * 100.0f) + 0.1f, 0.3f));
-    vertices.append(QVector3D((geoPos.x() * 100.0f) - 0.1f, (geoPos.y() * 100.0f) - 0.1f, 0.3f));
-    vertices.append(QVector3D((geoPos.x() * 100.0f) + 0.1f, (geoPos.y() * 100.0f) - 0.1f, 0.3f));
+    vertices.append(QVector3D((geoPos.x() * 100.0) + 0.1, (geoPos.y() * 100.0) + 0.1, 0.3f));
+    vertices.append(QVector3D((geoPos.x() * 100.0) - 0.1, (geoPos.y() * 100.0) + 0.1, 0.3f));
+    vertices.append(QVector3D((geoPos.x() * 100.0) - 0.1, (geoPos.y() * 100.0) - 0.1, 0.3f));
+    vertices.append(QVector3D((geoPos.x() * 100.0) + 0.1, (geoPos.y() * 100.0) - 0.1, 0.3f));
 }
 
-void GLPOIPoint::setupVertices(float scale)
+void GLPOIPoint::setupVertices(double scale)
 {
     vertices.clear();
     boost_xy_point geoPos = point->getGeoPosition();
 
-    if(scale > 1.0f)
-        scale = 1.0f;
+    if(scale > 1.0)
+        scale = 1.0;
 
-    vertices.append(QVector3D((geoPos.x() * 100.0f) + 0.1f * scale, (geoPos.y() * 100.0f) + 0.1f * scale, 0.3f));
-    vertices.append(QVector3D((geoPos.x() * 100.0f) - 0.1f * scale, (geoPos.y() * 100.0f) + 0.1f * scale, 0.3f));
-    vertices.append(QVector3D((geoPos.x() * 100.0f) - 0.1f * scale, (geoPos.y() * 100.0f) - 0.1f * scale, 0.3f));
-    vertices.append(QVector3D((geoPos.x() * 100.0f) + 0.1f * scale, (geoPos.y() * 100.0f) - 0.1f * scale, 0.3f));
+    vertices.append(QVector3D((geoPos.x() * 100.0) + 0.1 * scale, (geoPos.y() * 100.0) + 0.1 * scale, 0.3f));
+    vertices.append(QVector3D((geoPos.x() * 100.0) - 0.1 * scale, (geoPos.y() * 100.0) + 0.1 * scale, 0.3f));
+    vertices.append(QVector3D((geoPos.x() * 100.0) - 0.1 * scale, (geoPos.y() * 100.0) - 0.1 * scale, 0.3f));
+    vertices.append(QVector3D((geoPos.x() * 100.0) + 0.1 * scale, (geoPos.y() * 100.0) - 0.1 * scale, 0.3f));
 }
